Move ConversationSettings to its own file convsettings.py to avoid circular import issues
