+++ python/ta/typeagent/knowpro/secindex.py
@@ -1,10 +1,9 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-from typing import TYPE_CHECKING
-
 from ..aitools.embeddings import AsyncEmbeddingModel
 from ..aitools.vectorbase import TextEmbeddingIndexSettings
+from .convsettings import ConversationSettings
 from .interfaces import (
     IConversation,
     IConversationSecondaryIndexes,
@@ -21,9 +20,6 @@ from .reltermsindex import (
     build_related_terms_index,
 )
 from .timestampindex import TimestampToTextRangeIndex, build_timestamp_index
-
-if TYPE_CHECKING:
-    from .convutils import ConversationSettings
 
 
 class ConversationSecondaryIndexes(IConversationSecondaryIndexes):
@@ -80,7 +76,7 @@ async def build_secondary_indexes[
     TTermToSemanticRefIndex: ITermToSemanticRefIndex,
 ](
     conversation: IConversation[TMessage, TTermToSemanticRefIndex],
-    conversation_settings: "ConversationSettings",
+    conversation_settings: ConversationSettings,
 ) -> None:
     if conversation.secondary_indexes is None:
         storage_provider = await conversation_settings.get_storage_provider()
@@ -104,7 +100,7 @@ async def build_transient_secondary_indexes[
     TMessage: IMessage, TTermToSemanticRefIndex: ITermToSemanticRefIndex
 ](
     conversation: IConversation[TMessage, TTermToSemanticRefIndex],
-    conversation_settings: "ConversationSettings | None" = None,
+    conversation_settings: ConversationSettings | None = None,
 ) -> None:
     if conversation.secondary_indexes is None:
         # Try to get storage provider from conversation.settings first, then from parameter
