+++ python/ta/typeagent/podcasts/podcast_import.py
@@ -4,7 +4,7 @@
 import os
 import re
 
-from ..knowpro.convutils import ConversationSettings
+from ..knowpro.convsettings import ConversationSettings
 from ..knowpro.interfaces import Datetime
 from ..storage.utils import create_storage_provider
 from .podcast import Podcast, PodcastMessage, PodcastMessageMeta
